chore: split NSTemplateSet controller into multiple files
