+++ test/nstemplateset_assertion.go
@@ -69,9 +69,9 @@ func Provisioned() toolchainv1alpha1.Condition {
 
 func Provisioning() toolchainv1alpha1.Condition {
 	return toolchainv1alpha1.Condition{
-		Type:    toolchainv1alpha1.ConditionReady,
-		Status:  corev1.ConditionFalse,
-		Reason:  toolchainv1alpha1.NSTemplateSetProvisioningReason,
+		Type:   toolchainv1alpha1.ConditionReady,
+		Status: corev1.ConditionFalse,
+		Reason: toolchainv1alpha1.NSTemplateSetProvisioningReason,
 	}
 }
